socsec: Add --deterministic switch
This enables non-deterministic behaviour where the tool or its
dependencies can optionally no-longer read from pools of randomness
provided by the system. Instead, substitute in reads from /dev/zero or
equivalent.

This switch should only be used when invoked by the test suite. It
should not be invoked to sign production images. As such it is
suppressed from the --help output.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>