socsec: Introduce setuptools packaging
setup.py allows the scripts to more easily be installed in various
environments, and in the process of converting the scripts to setuptools
we fix the approach of discovering the OTP configuration files relative
to the current working directory.

Currently we configure setuptools to only install `otptool` and
`socsec`. `resocsec` is still in development so it is not installed for
now.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>