socsec: Ignore non-key data on PEM key files
The PEM format defines a one-line header consisting of '-----BEGIN',
where everything leading to that should be ignored.

This patch seeks the input key to that header, making it possible
to add text / comments to the PEM files without breaking
RSA.importKey.

For that, we're also changing the file descriptor from buffered
binary to text, which should help in case the PEM file comes
from operating systems with strange line-ending standards.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>