ast2600: Add support for non-overlapped SPL stack
Recent changes[1] to the Aspeed U-boot allows for the SPL image to
occupy the entire 64KB SRAM by not overlapping the verifiable
region with the SPL stack.

[1] - https://github.com/AspeedTech-BMC/u-boot/commit/a62cf9f

This patch adds a --stack_intersects_verification_region to socsec
(and equivalent switch to resocsec) that allows the signing of
images that occupy the entire 64KB verifiable region (including
the 512B signature), in case these types of images are being used.

The patch also adds an WARNING in case this setting is ommitted,
since that can break forward-compatibility once these new types
of images are the default in future SDKs.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>