otptool: fix cannot keep updating keys issue
If the ECC is enabled, it should be aware that the data needs to be 2DW
alignment for ECC calculation. When key_config is not 2DW aligned,
otptool will append default value to do ECC calculation.

Problem:
The odd dw offset default value is 0xffffffff. After ECC code has been
programmed, this odd dw offset value cannot be changed. In this case,
the last bit[13] is set to 0x1. So afterward keys is useless.

Solution:
Set no last bit for append value instead. When user wants to keep
updating keys, it needs to add "reserved" key type before new keys for
this ECC restriction.

Signed-off-by: Neal Liu <neal_liu@aspeedtech.com>
Change-Id: Icb9cfde11c5b2cbe7fcbe14591a3aa709658cdf3